0:34

+++ cifar10_ConvNet_v5_cat2/model.py
@@ -12,7 +12,7 @@ def _cat2(labels):
    table2 = tf.constant([0,0,1,1,1,1,1,1,0,0])
    A = tf.transpose(tf.stack([table1, table2], axis=0))
    one_hot = tf.one_hot(labels, 10, 1.0, 0.0, axis=-1)
-   labels_cat2 = tf.matmul(one_hot, A)
+   labels_cat2 = tf.matmul(one_hot, tf.to_float(A))
    labels_cat2 = tf.to_int64(tf.argmax(labels_cat2, axis=1))
    assert len(labels_cat2.get_shape()) == 1
    assert labels_cat2.get_shape()[0] == 100
